fix and add

+++ src/ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: marvin@42.fr <astachni>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/08 17:10:28 by marvin@42.f       #+#    #+#             */
-/*   Updated: 2022/11/09 13:09:05 by marvin@42.f      ###   ########.fr       */
+/*   Updated: 2022/11/09 16:31:33 by marvin@42.f      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ void	*ft_memchr(const void *memoryBlock, int searchedChar, size_t size)
 	i = 0;
 	while (i < size)
 	{
-		if (memoryBlock[i] == searchedChar)
+		if (((int)memoryBlock[i]) == searchedChar)
 		{
 			return (*memoryBlock[i]);
 		}
